Refactor to modular Flask project
